Refactor service_manager.rb and add test suite
- Extract service classes into lib/services/ directory
  - base_service.rb: shared helpers with new with_temp_config method
  - php_service.rb, mysql_service.rb, redis_service.rb
  - memcached_service.rb, nginx_service.rb, httpd_service.rb
- Reduce service_manager.rb from 1001 to 315 lines (orchestration only)
- Unify all comments to English
- Add Minitest test suite (22 tests, 39 assertions)
- Add Rakefile for rake test
- Add GitHub Actions CI workflow

Closes #16
